Changed P[ic] array to dynamic memory allocation to allow for longer array in combination with smaller dc. Setup for investigation of in-vivo 1hz for small dc.

+++ shotnoise.h
@@ -12,8 +12,8 @@
 
 #include <string.h>
 
-void getAlphas(float rate, float c_pre, float c_post, float *alphas); /* calculates alphas[1]=alphap and alphas[0]=alphad */
+void getAlphas(double rate, double c_pre, double c_post, double *alphas); /* calculates alphas[1]=alphap and alphas[0]=alphad */
 
-float updateWeight(float rho_old, float stepsize, float rate, float c_pre, float c_post); /* returns a new value for rho */
+double updateWeight(double rho_old, double stepsize, double rate, double c_pre, double c_post); /* returns a new value for rho */
 
 #endif /* SHOTNOISE_H */
